Refactor to use Function
