+++ spring-cloud-starter-stream-processor-httpclient/src/test/java/org/springframework/cloud/stream/app/httpclient/processor/HttpClientProcessorTests.java
@@ -16,11 +16,6 @@
 
 package org.springframework.cloud.stream.app.httpclient.processor;
 
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
-import static org.springframework.cloud.stream.test.matcher.MessageQueueMatcher.receivesPayloadThat;
-
 import java.util.Map;
 
 import org.hamcrest.Matchers;
@@ -34,7 +29,6 @@ import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfi
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
-import org.springframework.context.annotation.Import;
 import org.springframework.messaging.support.GenericMessage;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
@@ -44,6 +38,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+import static org.springframework.cloud.stream.test.matcher.MessageQueueMatcher.receivesPayloadThat;
 
 /**
  * Tests for Http Client Processor.
@@ -63,7 +62,7 @@ import org.springframework.web.bind.annotation.RestController;
 public abstract class HttpClientProcessorTests {
 
 	protected static final String BASE_URL =
-			"'http://localhost:' + @environment.getProperty('local.server.port')";
+		"'http://localhost:' + @environment.getProperty('local.server.port')";
 
 	@Autowired
 	protected Processor channels;
@@ -94,9 +93,9 @@ public abstract class HttpClientProcessorTests {
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
-			"httpclient.body={\"foo\":\"bar\"}",
-			"httpclient.httpMethod=POST"
+		"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
+		"httpclient.body={\"foo\":\"bar\"}",
+		"httpclient.httpMethod=POST"
 	})
 	public static class TestRequestPOSTTests extends HttpClientProcessorTests {
 
@@ -104,14 +103,14 @@ public abstract class HttpClientProcessorTests {
 		public void testRequest() {
 			channels.input().send(new GenericMessage<Object>("..."));
 			assertThat(messageCollector.forChannel(channels.output()),
-					receivesPayloadThat(Matchers.allOf(containsString("foo"), containsString("bar"))));
+				receivesPayloadThat(Matchers.allOf(containsString("foo"), containsString("bar"))));
 		}
 
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
-			"httpclient.httpMethod=POST"
+		"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
+		"httpclient.httpMethod=POST"
 	})
 	public static class TestRequestPOSTWithBodyExpressionTests extends HttpClientProcessorTests {
 
@@ -119,14 +118,14 @@ public abstract class HttpClientProcessorTests {
 		public void testRequest() {
 			channels.input().send(new GenericMessage<Object>("{\"foo\":\"bar\"}"));
 			assertThat(messageCollector.forChannel(channels.output()), receivesPayloadThat(
-					Matchers.allOf(containsString("Hello"), containsString("foo"), containsString("bar"))));
+				Matchers.allOf(containsString("Hello"), containsString("foo"), containsString("bar"))));
 		}
 
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/headers'",
-			"httpclient.headersExpression={Key1:'value1',Key2:'value2'}"
+		"httpclient.urlExpression= " + BASE_URL + " + '/headers'",
+		"httpclient.headersExpression={Key1:'value1',Key2:'value2'}"
 	})
 	public static class TestRequestWithHeadersTests extends HttpClientProcessorTests {
 
@@ -139,11 +138,11 @@ public abstract class HttpClientProcessorTests {
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
-			"httpclient.httpMethod=POST",
-			"httpclient.headersExpression={Accept:'application/octet-stream'}",
-			"httpclient.expectedResponseType=byte[]",
-			"spring.cloud.stream.bindings.output.contentType=application/octet-stream"
+		"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
+		"httpclient.httpMethod=POST",
+		"httpclient.headersExpression={Accept:'application/octet-stream'}",
+		"httpclient.expectedResponseType=byte[]",
+		"spring.cloud.stream.bindings.output.contentType=application/octet-stream"
 	})
 	public static class TestRequestWithReturnTypeTests extends HttpClientProcessorTests {
 
@@ -151,15 +150,15 @@ public abstract class HttpClientProcessorTests {
 		public void testRequest() {
 			channels.input().send(new GenericMessage<Object>("hello"));
 			assertThat(messageCollector.forChannel(channels.output()),
-					receivesPayloadThat(Matchers.isA(byte[].class)));
+				receivesPayloadThat(Matchers.isA(byte[].class)));
 		}
 
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
-			"httpclient.httpMethod=POST",
-			"httpclient.replyExpression=body.substring(3,8)"
+		"httpclient.urlExpression= " + BASE_URL + " + '/greet'",
+		"httpclient.httpMethod=POST",
+		"httpclient.replyExpression=body.substring(3,8)"
 	})
 	public static class TestRequestWithResultExtractorTests extends HttpClientProcessorTests {
 
@@ -172,8 +171,8 @@ public abstract class HttpClientProcessorTests {
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/json'",
-			"httpclient.httpMethod=POST", "httpclient.headersExpression={'Content-Type':'application/json'}"
+		"httpclient.urlExpression= " + BASE_URL + " + '/json'",
+		"httpclient.httpMethod=POST", "httpclient.headersExpression={'Content-Type':'application/json'}"
 
 	})
 	public static class TestRequestWithJsonPostTests extends HttpClientProcessorTests {
@@ -188,9 +187,9 @@ public abstract class HttpClientProcessorTests {
 	}
 
 	@TestPropertySource(properties = {
-			"httpclient.urlExpression= " + BASE_URL + " + '/json'",
-			"httpclient.httpMethodExpression=#jsonPath(payload,'$.myMethod')",
-			"httpclient.headersExpression={'Content-Type':'application/json'}"
+		"httpclient.urlExpression= " + BASE_URL + " + '/json'",
+		"httpclient.httpMethodExpression=#jsonPath(payload,'$.myMethod')",
+		"httpclient.headersExpression={'Content-Type':'application/json'}"
 	})
 	public static class TestRequestWithMethodExpressionTests extends HttpClientProcessorTests {
 
@@ -203,8 +202,9 @@ public abstract class HttpClientProcessorTests {
 
 	}
 
+	@SpringBootApplication(exclude = { SecurityAutoConfiguration.class, ManagementWebSecurityAutoConfiguration.class })
 	@RestController
-	public static class AdditionalController {
+	public static class HttpClientProcessorApplication {
 
 		@RequestMapping("/greet")
 		public String greet(@RequestBody(required = false) String who) {
@@ -223,12 +223,6 @@ public abstract class HttpClientProcessorTests {
 		public String json(@RequestBody Map<String, Object> request) {
 			return "id";
 		}
-
-	}
-
-	@SpringBootApplication(exclude = { SecurityAutoConfiguration.class, ManagementWebSecurityAutoConfiguration.class })
-	@Import(AdditionalController.class)
-	public static class HttpClientProcessorApplication {
 
 	}
